Add basic zombie AI movement
Uses the Pathfinder and A* PATH FINDER to calculate the minimun path
between the player and the IA and move the IA close to the player.

Some modifications using 3d box collider for the obstacles have been
added for the path fider to work properly.

A plane have been added for the path finder too.