Add PlayerCollision script and Hideout interaction
The PlayerCollision script is used for interactions with collisions
related with the Player, and other scripts can use his functions and
access some of his variables.

The ZombieDestinationSetter now pursues the player if the player is
NOT hidden. If the player IS hidden, the zombie moves to his default
location.