Add basic heath system for the player
The life scripts were based mostly from Rodrigo Lima "Vida" and
"BarraDeVida" scripts.

Since the player status need to be used for all different scenes of
the game, the current and max life/sanity were stored using a object
called "HealthManager" loaded on the _preload scene (remembering that
the _preload scene is loaded only once and before all scenes and
have objects that are active and can be used until the end of the game).

The player life bar can add or subtract values, update max health,
and always, at the start of the scene, it get the current status from
the 'HeathManager'.