Add a field of view for the zombie enemy
Now there is a script attached to the Zombie that is used to say if
the Player is inside the field of the vision of the Zombie.

This field of view can be modified in the inspector, changing the
radius and angle of the field.

Also, the Zombie cannot see the player if some 'Obstacle' is between
the player and the zombie, using a public LayerMask to test it.