Fix zombie field of view
Now the zombie really can see the player only throgth a circle arc
defined by the developer.

If the player runs from a certain distance ( defined by the dev in
the Inspector with the same radius of the vision circle arc ), then
the zombie can listen to the player and run following the player.