Add shoot and reload audio for player
Now the Audio Manager can be called from
any script to call the sound of the pistol
shooting, reloading or for the player taking
ammunition.

Basically the scene Teoi2 was created as a copy
of Rodrigo scene, and some scripts based on the
player shooting and reloading were modified to
call some AudioManager functions to play those
sounds.