Refactor Zombie Field of View behavior
As there is one object that have a 2Dcircle collider
attached to the player, the Zombie Field of View wasn't
working properly.

The version before this commit take a list of 2DCircleColliders
with the tag Player and it checks if the Length of the list is >= 1.
As the player now has more than one
2D CircleCollider with tag Player, the Zombie wrongly starts
to follow the player.

This commit now also checks the Name of the game object with the
collider. The name of the gameObject that has the collider needs
to be "Player 1".